Multi-Slot Breeding Permissions
+ Add: Quantity field for breeding permissions.
Now players & admins can create multiple permissions in one. Great for fast permission granting! The default max is set to 10. If you wish to change this, change in the applicable blades and the breeding service. (Just see commit changes if you're unsure where)
+ Add: When transferring a portion of the remaining quantity, the service will create a NEW breeding permission and deduct the transferred quantity from the old permission.

If you are NOT interested in having multi-slot breeding perms. You are free to pull the commit before this one for a pure breeding-permissions ext for v3!